+++ src/main/java/com/lucendar/common/types/KV.java
@@ -1,0 +1,41 @@
+package com.lucendar.common.types;
+
+import java.util.StringJoiner;
+
+/**
+ * Key value tuple
+ */
+public class KV {
+
+    private final String key;
+    private final Object value;
+
+    public KV(String key, Object value) {
+        this.key = key;
+        this.value = value;
+    }
+
+    public String key() {
+        return key;
+    }
+
+    public Object value() {
+        return value;
+    }
+
+    public String getKey() {
+        return key;
+    }
+
+    public Object getValue() {
+        return value;
+    }
+
+    @Override
+    public String toString() {
+        return new StringJoiner(", ", KV.class.getSimpleName() + "[", "]")
+                .add("key='" + key + "'")
+                .add("value=" + value)
+                .toString();
+    }
+}